DS2-245: Support uuid in firestore-service (#31)
* DS2-245: Support uuid in firestore-service

* Do not duplicate literal, use named constant

* Update __init__.py

* Use an Enum to group special ID/UUID values

* Fix test `test_set_history_doc_refs_existing_by_url_with_wrong_dealroom_uuid`

Looking for a document by ID (UUID) but matching other documents by final_url (which happens when final_url is in payload) means that not all documents will have field "dealroom_id" ("dealroom_uuid"). In that case, the check for in progress or deleted should not fail. Therefore, we access the key using `.get(key)` instead of `[key]`.

* Rename Code to StatusCode

* Refactor silly logic